requirements.txt
PyQt6>=6.4.2
PyPDF2>=3.0.0
python-docx>=0.8.11
ebooklib>=0.18
pydub>=0.25.1
ffmpeg-python>=0.2.0
pyttsx3>=2.90
edge-tts>=7.2.3
requests>=2.28.0
piper-tts>=1.2.0
numpy>=1.21.0
